frontend/backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.1
marshmallow==3.20.2
flask-marshmallow==1.0.0
marshmallow-sqlalchemy==1.0.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
werkzeug==3.0.1
pillow==10.2.0
python-magic==0.4.27
flask-jwt-extended==4.6.0
flask-mail==0.9.1
itsdangerous==2.1.2
bcrypt==4.1.2 